+++ llarp/link/link_manager.hpp
@@ -254,14 +254,12 @@ namespace llarp
         // This will try to connect to *up to* num_conns routers, but will not
         // check if we already have a connection to any of the random set, as making
         // that thread safe would be slow...I think.
-        void connect_to_random(int num_conns, bool client_only = false);
+        void connect_to_random(size_t num_conns, bool client_only = false);
 
         /// always maintain this many client connections to other routers
         int client_router_connections = 4;
 
        private:
-
-
         // DHT messages
         void handle_find_name(std::string_view body, std::function<void(std::string)> respond);      // relay
         void handle_find_intro(std::string_view body, std::function<void(std::string)> respond);     // relay
